requirements.txt
# Kafka
confluent-kafka==2.1.1

# Spark
pyspark==3.3.0

# Airflow
apache-airflow==2.5.0
apache-airflow-providers-snowflake==4.0.0

# Storage
boto3==1.26.0
minio==7.1.15

# Data Sources
yfinance==0.2.18
alpha-vantage==2.3.1

# Data Processing
pandas==1.5.3
numpy==1.24.2
pyarrow==11.0.0

# Snowflake
snowflake-connector-python==3.0.4
snowflake-sqlalchemy==1.4.7

# dbt
dbt-snowflake==1.7.0
dbt-core==1.7.0

# Utilities
python-dotenv==1.0.0
requests==2.28.2
